add createZagHook function
